+++ packages/contracts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.72](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@0.0.71...@injectivelabs/contracts@0.0.72) (2021-09-19)
+
+**Note:** Version bump only for package @injectivelabs/contracts
+
+
+
+
+
 ## [0.0.71](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@0.0.70...@injectivelabs/contracts@0.0.71) (2021-09-18)